Added a checkbox that provides the user the option to hide the splash screen next time.
Implemented the option for the user to hide the splash screen next time
he connects. The preferences are saved using cookie.

Some function should be moved to a utility class in the future as they
have really nothing specific to the splash screen logic and can be
reused in an number of occasion.

Also changed  the css layout of the page to make it more responsive and
maintain some spaces between the different text elements.