Syle changes
Made the bottom margin smaller and make the text box take a bigger
height of the screen. The splash had 2 defaults: 1. too much space
between text and button with the window has a large height value, and
button and radio box tended to overlap the text too fast when height was
small.