Splash screen (#6)
* Added template for splash screen.

Added template for splash screen.

* Modified the style of the splash.

Made it vertically centered

* Created the configuration for the splash screen.

Added splash screen settings and a splash tool.

* Corrected typo in configuration panel

Corrected typo in configuration panel

* Removed the image field

The image field is not required since the splash screen text control is
a rich text editor which allow the user to type text, add links, and
reference image through links.

* Updated the read me

Fixed syle issues (title style not used because the read me is missing
carriage return after title) and added a link to the documentation about
adding parameters to a configurable template.

* Configurable splash screen

The splash screen get the following information from the user
configuration:
- Size
- Content
- Background Color

It has a close button and the content of the div is set to be scrollable
vertically if the text overflows.

* Fixed some display bugs for the splash screen.

Scrolling was not working, and the close button seems a bit out of
place.  Made the positionning a bit easier to understand. Still need to
move css properties for the close button from javascript to CSS.

* Moved splash screen element properties from javascript to css.

Declare a common class shared between the splash screen text and the
button container and then indiviual class for the properties where those
things are different. Overall, it help giving those element a consistent
positioning.

* Removed comments from the splash module code.

Removed comments from the splash module code.

* Modified the splash screen to attach it to an div of index.html.

Use a div that is always there and stopped destroying the container div.

* Improved splash screen accessibility.

Made the splash screen accessible using tabs and gave it content that
can be read by text reader.

* Added a checkbox that provides the user the option to hide the splash screen next time.

Implemented the option for the user to hide the splash screen next time
he connects. The preferences are saved using cookie.

Some function should be moved to a utility class in the future as they
have really nothing specific to the splash screen logic and can be
reused in an number of occasion.

Also changed  the css layout of the page to make it more responsive and
maintain some spaces between the different text elements.

* Code clean up

Added code comments for the splash widget.

* Moved some utility function from the splash module to a utility module.

Moved some utility function from the splash module to a utility module.

* Remove the need to instanciate the wabutils modules.

Since it provides utility function, this module should expose "static"
method. There is no need for instanciating objects from the client side.

* Code clean up.

Code clean up.

* Corrected bug.

The value of the checkbox about not showing the splash screen was not
refreched if the user was: updating the value of the splash screen
checkbox, and then chose to display it through the keyboard navigation.

* Added a separate ratio for width ratio and height ratio for the splash screen + bug fix on checkbox

Added a separate ratio for width ratio and height ratio for the splash
screen + bug fix on checkbox

* Improved the tab index for accessing the spash screen content.

Improved the tab index for accessing the spash screen content. Works for
the textbox and the button, but the checkbox does not get highlighted.

* Modified the checkbox.

Replaced dojo check box by a regular text box so that it follows the
same tab tab navigation display by default than any other component.
Also made the checkbox bigger.

* Syle changes

Made the bottom margin smaller and make the text box take a bigger
height of the screen. The splash had 2 defaults: 1. too much space
between text and button with the window has a large height value, and
button and radio box tended to overlap the text too fast when height was
small.

* Added aknowledgement

Added aknowledgements to MCE, who paid for the developement of the
splash screen widget.

* Updated the splash screen code to create the content only the first time it needs to be displayed to the user.

No need to create the HTML content of the splash  screen before the user
asks for it. It is extra processing that may not be used.

* Bugs Correction

Corrected some bugs in the splash screen text being undefined and
modified the default behaviour of the splash screen.

* Added code to center the text in the spash screen

When not enough text is not provided, the text should be vertically
centered within the text div. This is visually more appealing.

* Added a shortcut to access the splash screen.

Useful if the user click or get out if the screen and want to get back
to it.

* Improved the positionning of the text box

Improved the positionning of the text box

* Added a top margin from the splash screen

Added a top margin from the splash screen

* Updated aknowledgments to MCE.

Updated aknowledgments to MCE to take into account their comments.