Added catch exception when the feature list does not manage to process a layer
When the FeatureList widget fails to read a service that contains
multiple layers the code fails and the application crashes abrubtly.
This code changes simply wraps the logic in a try/catch exception. If
there is problem with a layer, a warning is reported in the console and
the application keep working with the layer that it can support.