Fixed case related issues in paths. (#7)
Some of the widget have case issues when pointing to external modules or
files. This does not mean much on Windows where file name is case
insensitive, but it matter when deploying the app on a Linux machine. On
linux, there is still an issue with loading the image
images/someFilters,png that should be fix at a latter time.