Added a minimal extent for the seach to zoom at.
When zooming on a point the search widget, the widget has a tendency to
go beyond the maximum zoom level supported by ArcGIS Online base maps.
This proposed fix sets a minimum scale for the search widget, to keep it
visible with the basemap, when the result set does not have an extent.