Added zoom and navigation fix.
For search widget. If zooming on point, the widget zoom beyond the
basemap max zoom. This code contains a fix for that.

For Layers widget, a bug fix to let the user check / uncheck boxes and
therefore toogle layers via the keyboard. Was not working in Chrome and
IE11 before.